Keep lattice dispatch tests out of math lane
